requirements.txt
tf-keras==2.17.0
transformers==4.55.0
huggingface-hub==0.34.3
accelerate==1.10.0
sentencepiece==0.2.0
bitsandbytes==0.46.1
numpy==1.26.4
pandas==1.5.3
pillow==11.3.0
librosa==0.10.2
ipython==8.25.0
cloudpickle==2.2.1
dill==0.3.8
lz4==4.3.2
flask==3.1.1
mlflow==2.21.3
boto3==1.34.69
google-cloud-storage==2.10.0
pyopenssl==24.0.0
defusedxml==0.7.1
configparser==5.2.0